+++ stackit/internal/services/mongodbflex/user/datasource_test.go
@@ -47,7 +47,7 @@ func TestMapDataSourceFields(t *testing.T) {
 					Username: utils.Ptr("username"),
 					Database: utils.Ptr("database"),
 					Host:     utils.Ptr("host"),
-					Port:     utils.Ptr(int32(1234)),
+					Port:     utils.Ptr(int64(1234)),
 				},
 			},
 			DataSourceModel{
@@ -76,7 +76,7 @@ func TestMapDataSourceFields(t *testing.T) {
 					Username: nil,
 					Database: nil,
 					Host:     nil,
-					Port:     utils.Ptr(int32(2123456789)),
+					Port:     utils.Ptr(int64(2123456789)),
 				},
 			},
 			DataSourceModel{
